+++ src/filters.php
@@ -6,19 +6,6 @@ use Roots\Sage\Template;
 use Roots\Sage\Template\Wrapper;
 
 /**
- * Determine which pages should NOT display the sidebar
- * @link https://codex.wordpress.org/Conditional_Tags
- */
-add_filter('sage/display_sidebar', function ($display) {
-    // The sidebar will NOT be displayed if ANY of the following return true
-    return $display ? !in_array(true, [
-        is_404(),
-        is_front_page(),
-        is_page_template('templates/template-custom.php'),
-    ]) : $display;
-});
-
-/**
  * Add <body> classes
  */
 add_filter('body_class', function (array $classes) {