+++ src/helpers.php
@@ -35,7 +35,7 @@ function asset_path($filename)
 function display_sidebar()
 {
     static $display;
-    isset($display) || $display = apply_filters('sage/display_sidebar', true);
+    isset($display) || $display = apply_filters('sage/display_sidebar', false);
     return $display;
 }